test: reorganize tests into action/behavior/ by condition category
- Consolidate scattered test files from action/tiles/, action/play_card/, action/card_effects/, and game/cards/ into action/behavior/
- One file per BehaviorCondition category (basic_resource, production, tile_placement, global_parameter, etc.)
- Add new tests for colony, tile_modification, misc, and effect condition types
- Add shared test helpers in helpers_test.go for applyOutputs, assertResources, etc.
- Update backend/test/CLAUDE.md to document the new test organization
