+++ backend/test/action/behavior/card_storage_test.go
@@ -1,4 +1,4 @@
-package card_effects_test
+package behavior_test
 
 import (
 	"context"
@@ -10,6 +10,10 @@ import (
 	"terraforming-mars-backend/internal/game/shared"
 	"terraforming-mars-backend/test/testutil"
 )
+
+// ============================================================================
+// Card resource tests
+// ============================================================================
 
 // --- CEO's Favorite Project (149) ---
 // "Add 1 resource to a card with at least 1 resource on it"
@@ -137,7 +141,7 @@ func TestCardResource_CEOsFavoriteProject_AddsToMicrobeCard(t *testing.T) {
 }
 
 // --- Corroder Suits (219) ---
-// "Increase your M€ production 2 steps. Add 1 resource to **any venus card**."
+// "Increase your M$ production 2 steps. Add 1 resource to **any venus card**."
 
 func TestCardResource_CorroderSuits_AddsToVenusCard(t *testing.T) {
 	broadcaster := testutil.NewMockBroadcaster()
@@ -406,7 +410,6 @@ func TestCardResource_FailsWhenTargetHasNoStorage(t *testing.T) {
 
 // --- Hydrogen To Venus (231) ---
 // "Raise Venus 1 step. Add 1 floater to a venus card for each Jovian tag you have."
-// When no venus card exists to receive floaters, the card should still play successfully.
 
 func TestCardResource_AnyCardTarget_SkipsWhenNoTargetCard(t *testing.T) {
 	broadcaster := testutil.NewMockBroadcaster()
@@ -479,4 +482,286 @@ func TestCardResource_AnyCardTarget_SkipsWhenNoTargetCard(t *testing.T) {
 
 	// Verify card was removed from hand
 	testutil.AssertEqual(t, false, p.Hand().HasCard("card-hydrogen-to-venus"), "Card should be removed from hand")
+}
+
+// ============================================================================
+// Card steal tests
+// ============================================================================
+
+func TestPredatorsStealAnimalFromOtherPlayer(t *testing.T) {
+	testGame, repo, cardRegistry, playerID, otherPlayerID := testutil.SetupTwoPlayerGame(t)
+	logger := testutil.TestLogger()
+
+	p, _ := testGame.GetPlayer(playerID)
+	other, _ := testGame.GetPlayer(otherPlayerID)
+
+	predatorsID := "test-predators"
+	p.PlayedCards().AddCard(predatorsID, "Predators", "active", []string{"animal"})
+	p.Resources().AddToStorage(predatorsID, 0)
+
+	targetCardID := "test-birds"
+	other.PlayedCards().AddCard(targetCardID, "Birds", "active", []string{"animal"})
+	other.Resources().AddToStorage(targetCardID, 3)
+
+	predatorsBehavior := shared.CardBehavior{
+		Triggers: []shared.Trigger{{Type: shared.TriggerTypeManual}},
+		Outputs: []shared.BehaviorCondition{
+			shared.NewCardStorageCondition(shared.ResourceAnimal, 1, "steal-from-any-card"),
+		},
+	}
+	p.Actions().SetActions([]shared.CardAction{
+		{
+			CardID:        predatorsID,
+			CardName:      "Predators",
+			BehaviorIndex: 0,
+			Behavior:      predatorsBehavior,
+		},
+	})
+
+	useAction := cardAction.NewUseCardActionAction(repo, cardRegistry, nil, logger)
+	ctx := context.Background()
+
+	err := useAction.Execute(ctx, testGame.ID(), playerID, predatorsID, 0, nil, []string{predatorsID}, nil, &targetCardID, nil, nil, nil)
+	testutil.AssertNoError(t, err, "Predators steal action should succeed")
+
+	testutil.AssertEqual(t, 2, other.Resources().GetCardStorage(targetCardID), "Target card should have 2 animals after steal")
+	testutil.AssertEqual(t, 1, p.Resources().GetCardStorage(predatorsID), "Predators card should have 1 animal after steal")
+}
+
+func TestPredatorsStealFromOwnCard(t *testing.T) {
+	testGame, repo, cardRegistry, playerID, _ := testutil.SetupTwoPlayerGame(t)
+	logger := testutil.TestLogger()
+
+	p, _ := testGame.GetPlayer(playerID)
+
+	predatorsID := "test-predators"
+	p.PlayedCards().AddCard(predatorsID, "Predators", "active", []string{"animal"})
+	p.Resources().AddToStorage(predatorsID, 0)
+
+	targetCardID := "test-fish"
+	p.PlayedCards().AddCard(targetCardID, "Fish", "active", []string{"animal"})
+	p.Resources().AddToStorage(targetCardID, 2)
+
+	predatorsBehavior := shared.CardBehavior{
+		Triggers: []shared.Trigger{{Type: shared.TriggerTypeManual}},
+		Outputs: []shared.BehaviorCondition{
+			shared.NewCardStorageCondition(shared.ResourceAnimal, 1, "steal-from-any-card"),
+		},
+	}
+	p.Actions().SetActions([]shared.CardAction{
+		{
+			CardID:        predatorsID,
+			CardName:      "Predators",
+			BehaviorIndex: 0,
+			Behavior:      predatorsBehavior,
+		},
+	})
+
+	useAction := cardAction.NewUseCardActionAction(repo, cardRegistry, nil, logger)
+	ctx := context.Background()
+
+	err := useAction.Execute(ctx, testGame.ID(), playerID, predatorsID, 0, nil, []string{predatorsID}, nil, &targetCardID, nil, nil, nil)
+	testutil.AssertNoError(t, err, "Predators steal from own card should succeed")
+
+	testutil.AssertEqual(t, 1, p.Resources().GetCardStorage(targetCardID), "Target card should have 1 animal after steal")
+	testutil.AssertEqual(t, 1, p.Resources().GetCardStorage(predatorsID), "Predators card should have 1 animal after steal")
+}
+
+func TestPredatorsRejectsWithNoSourceCard(t *testing.T) {
+	testGame, repo, cardRegistry, playerID, _ := testutil.SetupTwoPlayerGame(t)
+	logger := testutil.TestLogger()
+
+	p, _ := testGame.GetPlayer(playerID)
+
+	predatorsID := "test-predators"
+	p.PlayedCards().AddCard(predatorsID, "Predators", "active", []string{"animal"})
+	p.Resources().AddToStorage(predatorsID, 0)
+
+	predatorsBehavior := shared.CardBehavior{
+		Triggers: []shared.Trigger{{Type: shared.TriggerTypeManual}},
+		Outputs: []shared.BehaviorCondition{
+			shared.NewCardStorageCondition(shared.ResourceAnimal, 1, "steal-from-any-card"),
+		},
+	}
+	p.Actions().SetActions([]shared.CardAction{
+		{
+			CardID:        predatorsID,
+			CardName:      "Predators",
+			BehaviorIndex: 0,
+			Behavior:      predatorsBehavior,
+		},
+	})
+
+	useAction := cardAction.NewUseCardActionAction(repo, cardRegistry, nil, logger)
+	ctx := context.Background()
+
+	err := useAction.Execute(ctx, testGame.ID(), playerID, predatorsID, 0, nil, []string{predatorsID}, nil, nil, nil, nil, nil)
+	testutil.AssertError(t, err, "Predators without source card should be rejected")
+	testutil.AssertEqual(t, 0, p.Resources().GetCardStorage(predatorsID), "Predators card should have 0 animals when no source card specified")
+}
+
+func TestPredatorsStealFromCardWithZeroAnimals(t *testing.T) {
+	testGame, repo, cardRegistry, playerID, otherPlayerID := testutil.SetupTwoPlayerGame(t)
+	logger := testutil.TestLogger()
+
+	p, _ := testGame.GetPlayer(playerID)
+	other, _ := testGame.GetPlayer(otherPlayerID)
+
+	predatorsID := "test-predators"
+	p.PlayedCards().AddCard(predatorsID, "Predators", "active", []string{"animal"})
+	p.Resources().AddToStorage(predatorsID, 0)
+
+	targetCardID := "test-birds"
+	other.PlayedCards().AddCard(targetCardID, "Birds", "active", []string{"animal"})
+	other.Resources().AddToStorage(targetCardID, 0)
+
+	predatorsBehavior := shared.CardBehavior{
+		Triggers: []shared.Trigger{{Type: shared.TriggerTypeManual}},
+		Outputs: []shared.BehaviorCondition{
+			shared.NewCardStorageCondition(shared.ResourceAnimal, 1, "steal-from-any-card"),
+		},
+	}
+	p.Actions().SetActions([]shared.CardAction{
+		{
+			CardID:        predatorsID,
+			CardName:      "Predators",
+			BehaviorIndex: 0,
+			Behavior:      predatorsBehavior,
+		},
+	})
+
+	useAction := cardAction.NewUseCardActionAction(repo, cardRegistry, nil, logger)
+	ctx := context.Background()
+
+	err := useAction.Execute(ctx, testGame.ID(), playerID, predatorsID, 0, nil, []string{predatorsID}, nil, &targetCardID, nil, nil, nil)
+	testutil.AssertNoError(t, err, "Predators steal from empty card should succeed")
+
+	testutil.AssertEqual(t, 0, other.Resources().GetCardStorage(targetCardID), "Target card should still have 0 animals")
+	testutil.AssertEqual(t, 0, p.Resources().GetCardStorage(predatorsID), "Predators card should have 0 animals when source had none")
+}
+
+func TestAntsStealMicrobeFromOtherPlayer(t *testing.T) {
+	testGame, repo, cardRegistry, playerID, otherPlayerID := testutil.SetupTwoPlayerGame(t)
+	logger := testutil.TestLogger()
+
+	p, _ := testGame.GetPlayer(playerID)
+	other, _ := testGame.GetPlayer(otherPlayerID)
+
+	antsID := "test-ants"
+	p.PlayedCards().AddCard(antsID, "Ants", "active", []string{"microbe"})
+	p.Resources().AddToStorage(antsID, 0)
+
+	targetCardID := "test-decomposers"
+	other.PlayedCards().AddCard(targetCardID, "Decomposers", "active", []string{"microbe"})
+	other.Resources().AddToStorage(targetCardID, 5)
+
+	antsBehavior := shared.CardBehavior{
+		Triggers: []shared.Trigger{{Type: shared.TriggerTypeManual}},
+		Outputs: []shared.BehaviorCondition{
+			shared.NewCardStorageCondition(shared.ResourceMicrobe, 1, "steal-from-any-card"),
+		},
+	}
+	p.Actions().SetActions([]shared.CardAction{
+		{
+			CardID:        antsID,
+			CardName:      "Ants",
+			BehaviorIndex: 0,
+			Behavior:      antsBehavior,
+		},
+	})
+
+	useAction := cardAction.NewUseCardActionAction(repo, cardRegistry, nil, logger)
+	ctx := context.Background()
+
+	err := useAction.Execute(ctx, testGame.ID(), playerID, antsID, 0, nil, []string{antsID}, nil, &targetCardID, nil, nil, nil)
+	testutil.AssertNoError(t, err, "Ants steal action should succeed")
+
+	testutil.AssertEqual(t, 4, other.Resources().GetCardStorage(targetCardID), "Target card should have 4 microbes after steal")
+	testutil.AssertEqual(t, 1, p.Resources().GetCardStorage(antsID), "Ants card should have 1 microbe after steal")
+}
+
+func TestAntsRejectsWithNoSourceCard(t *testing.T) {
+	testGame, repo, cardRegistry, playerID, _ := testutil.SetupTwoPlayerGame(t)
+	logger := testutil.TestLogger()
+
+	p, _ := testGame.GetPlayer(playerID)
+
+	antsID := "test-ants"
+	p.PlayedCards().AddCard(antsID, "Ants", "active", []string{"microbe"})
+	p.Resources().AddToStorage(antsID, 0)
+
+	antsBehavior := shared.CardBehavior{
+		Triggers: []shared.Trigger{{Type: shared.TriggerTypeManual}},
+		Outputs: []shared.BehaviorCondition{
+			shared.NewCardStorageCondition(shared.ResourceMicrobe, 1, "steal-from-any-card"),
+		},
+	}
+	p.Actions().SetActions([]shared.CardAction{
+		{
+			CardID:        antsID,
+			CardName:      "Ants",
+			BehaviorIndex: 0,
+			Behavior:      antsBehavior,
+		},
+	})
+
+	useAction := cardAction.NewUseCardActionAction(repo, cardRegistry, nil, logger)
+	ctx := context.Background()
+
+	err := useAction.Execute(ctx, testGame.ID(), playerID, antsID, 0, nil, []string{antsID}, nil, nil, nil, nil, nil)
+	testutil.AssertError(t, err, "Ants without source card should be rejected")
+	testutil.AssertEqual(t, 0, p.Resources().GetCardStorage(antsID), "Ants card should have 0 microbes when no source card specified")
+}
+
+func TestCardStorage_VariableAmountInput(t *testing.T) {
+	testGame, _, _, playerID, _ := testutil.SetupTwoPlayerGame(t)
+
+	p, _ := testGame.GetPlayer(playerID)
+
+	cardID := "test-microbe-card"
+	p.PlayedCards().AddCard(cardID, "Test Microbe Card", "active", []string{"microbe"})
+	p.Resources().AddToStorage(cardID, 5)
+
+	input := &shared.CardStorageCondition{
+		ConditionBase:  shared.ConditionBase{ResourceType: shared.ResourceMicrobe, Amount: 1, Target: "self-card"},
+		VariableAmount: true,
+	}
+
+	applier := gamecards.NewBehaviorApplier(p, testGame, "test", testutil.TestLogger()).
+		WithSelectedAmount(2).
+		WithSourceCardID(cardID)
+	err := applier.ApplyInputs(context.Background(), []shared.BehaviorCondition{input})
+	testutil.AssertNoError(t, err, "ApplyInputs should succeed")
+
+	testutil.AssertEqual(t, 3, p.Resources().GetCardStorage(cardID), "Should have 3 microbes remaining after spending 2")
+}
+
+func TestCardStorage_StealFromCardWithZeroResources(t *testing.T) {
+	testGame, _, _, playerID, otherPlayerID := testutil.SetupTwoPlayerGame(t)
+
+	p, _ := testGame.GetPlayer(playerID)
+	other, _ := testGame.GetPlayer(otherPlayerID)
+
+	// Player 2 has a card with animal storage but 0 animals
+	otherCardID := "test-empty-animals"
+	other.PlayedCards().AddCard(otherCardID, "Empty Animals", "active", []string{"animal"})
+	other.Resources().AddToStorage(otherCardID, 0)
+
+	// Player 1 has a card with animal storage and some animals
+	selfCardID := "test-predators"
+	p.PlayedCards().AddCard(selfCardID, "Predators", "active", []string{"animal"})
+	p.Resources().AddToStorage(selfCardID, 3)
+
+	output := shared.NewCardStorageCondition(shared.ResourceAnimal, 2, "steal-from-any-card")
+
+	applyOutputsWithOptions(t, p, testGame, applyOptions{
+		sourceCardID:      selfCardID,
+		stealSourceCardID: otherCardID,
+	}, output)
+
+	// Nothing to steal from empty card
+	testutil.AssertEqual(t, 0, other.Resources().GetCardStorage(otherCardID),
+		"Other player's card should still have 0 animals")
+	testutil.AssertEqual(t, 3, p.Resources().GetCardStorage(selfCardID),
+		"Player's card should still have original 3 animals (nothing stolen)")
 }
