--- backend/test/action/tiles/tile_adjacency_restriction_test.go
@@ -1,603 +1,0 @@
-package tiles_test
-
-import (
-	"context"
-	"testing"
-
-	"terraforming-mars-backend/internal/action"
-	cardAction "terraforming-mars-backend/internal/action/card"
-	"terraforming-mars-backend/internal/cards"
-	"terraforming-mars-backend/internal/game/board"
-	gamecards "terraforming-mars-backend/internal/game/cards"
-	"terraforming-mars-backend/internal/game/player"
-	"terraforming-mars-backend/internal/game/shared"
-	"terraforming-mars-backend/test/testutil"
-)
-
-// --- Urbanized Area (120) ---
-// "Place a city tile adjacent to at least 2 other city tiles."
-
-func TestUrbanizedArea_CityAdjacentTo2Cities(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, repo := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	logger := testutil.TestLogger()
-	ctx := context.Background()
-
-	minAdj := 2
-	urbanizedArea := gamecards.Card{
-		ID:   "card-urbanized-area",
-		Name: "Urbanized Area",
-		Type: gamecards.CardTypeAutomated,
-		Cost: 10,
-		Tags: []shared.CardTag{shared.TagBuilding, shared.TagCity},
-		Behaviors: []shared.CardBehavior{
-			{
-				Triggers: []shared.Trigger{{Type: shared.TriggerTypeAuto}},
-				Outputs: []shared.BehaviorCondition{
-					shared.NewProductionCondition(shared.ResourceCreditProduction, 2, "self-player"),
-					shared.NewProductionCondition(shared.ResourceEnergyProduction, -1, "self-player"),
-					&shared.TilePlacementCondition{
-						ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceCityPlacement, Amount: 1, Target: "none"},
-						TileRestrictions: &shared.TileRestrictions{
-							AdjacentToType:    "city",
-							MinAdjacentOfType: &minAdj,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{urbanizedArea})
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{
-		shared.ResourceCredit: 100,
-	})
-	p.Resources().AddProduction(map[shared.ResourceType]int{
-		shared.ResourceEnergyProduction: 1,
-	})
-	p.Hand().AddCard("card-urbanized-area")
-
-	// Place 2 adjacent city tiles: (0,0,0) and (1,0,-1) are neighbors
-	city1 := shared.HexPosition{Q: 0, R: 0, S: 0}
-	city2 := shared.HexPosition{Q: 1, R: 0, S: -1}
-
-	err := testGame.Board().UpdateTileOccupancy(ctx, city1,
-		board.TileOccupant{Type: shared.ResourceCityTile}, "other-player")
-	testutil.AssertNoError(t, err, "placing city1")
-
-	err = testGame.Board().UpdateTileOccupancy(ctx, city2,
-		board.TileOccupant{Type: shared.ResourceCityTile}, "other-player")
-	testutil.AssertNoError(t, err, "placing city2")
-
-	playCardAction := cardAction.NewPlayCardAction(repo, cardRegistry, nil, logger)
-	payment := cardAction.PaymentRequest{Credits: 10}
-	err = playCardAction.Execute(ctx, testGame.ID(), p.ID(), "card-urbanized-area", payment, nil, nil, nil, nil)
-	testutil.AssertNoError(t, err, "Urbanized Area should play successfully")
-
-	selection := testGame.GetPendingTileSelection(p.ID())
-	testutil.AssertTrue(t, selection != nil, "Should have pending tile selection")
-	testutil.AssertEqual(t, "city", selection.TileType, "Pending tile type should be city")
-
-	// The only hexes adjacent to BOTH (0,0,0) and (1,0,-1) are:
-	// (0,1,-1) — neighbor of both via cube coords
-	// (1,-1,0) — neighbor of both via cube coords
-	// These should be the available hexes (both are land tiles, not ocean/tagged)
-	expectedHexes := map[string]bool{
-		shared.HexPosition{Q: 0, R: 1, S: -1}.String(): true, // adjacent to both cities
-		shared.HexPosition{Q: 1, R: -1, S: 0}.String(): true, // adjacent to both cities
-	}
-
-	for _, hex := range selection.AvailableHexes {
-		testutil.AssertTrue(t, expectedHexes[hex],
-			"Hex "+hex+" should be adjacent to 2+ cities")
-	}
-	testutil.AssertTrue(t, len(selection.AvailableHexes) > 0,
-		"Should have at least one available hex adjacent to 2 cities")
-}
-
-func TestUrbanizedArea_NoCitiesYield0AvailableHexes(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, _ := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	ctx := context.Background()
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	// No cities placed — should have 0 available hexes with adjacentToType: "city", minAdjacentOfType: 2
-	minAdj := 2
-	count := testGame.CountAvailableHexesForTile("city", p.ID(), &shared.TileRestrictions{
-		AdjacentToType:    "city",
-		MinAdjacentOfType: &minAdj,
-	})
-	testutil.AssertEqual(t, 0, count, "No hexes should be available when no cities exist")
-}
-
-func TestUrbanizedArea_OnlyOneCityYield0AvailableHexes(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, _ := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	ctx := context.Background()
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	// Place only 1 city — no hex can be adjacent to 2 cities
-	city1 := shared.HexPosition{Q: 0, R: 0, S: 0}
-	err := testGame.Board().UpdateTileOccupancy(ctx, city1,
-		board.TileOccupant{Type: shared.ResourceCityTile}, p.ID())
-	testutil.AssertNoError(t, err, "placing city")
-
-	minAdj := 2
-	count := testGame.CountAvailableHexesForTile("city", p.ID(), &shared.TileRestrictions{
-		AdjacentToType:    "city",
-		MinAdjacentOfType: &minAdj,
-	})
-	testutil.AssertEqual(t, 0, count, "No hexes should be available when only 1 city exists")
-}
-
-// --- Ecological Zone (128) ---
-// "Place this tile adjacent to any greenery tile."
-
-func TestEcologicalZone_GreeneryAdjacentToGreenery(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, repo := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	logger := testutil.TestLogger()
-	ctx := context.Background()
-
-	ecologicalZone := gamecards.Card{
-		ID:   "card-ecological-zone",
-		Name: "Ecological Zone",
-		Type: gamecards.CardTypeActive,
-		Cost: 12,
-		Tags: []shared.CardTag{shared.TagAnimal, shared.TagPlant},
-		Behaviors: []shared.CardBehavior{
-			{
-				Triggers: []shared.Trigger{{Type: shared.TriggerTypeAuto}},
-				Outputs: []shared.BehaviorCondition{
-					shared.NewCardStorageCondition(shared.ResourceAnimal, 2, "self-card"),
-					&shared.TilePlacementCondition{
-						ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceGreeneryPlacement, Amount: 1, Target: "none"},
-						TileRestrictions: &shared.TileRestrictions{
-							AdjacentToType: "greenery",
-						},
-					},
-				},
-			},
-		},
-		ResourceStorage: &gamecards.ResourceStorage{
-			Type:     shared.ResourceAnimal,
-			Starting: 0,
-		},
-	}
-
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{ecologicalZone})
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{
-		shared.ResourceCredit: 100,
-	})
-	p.Hand().AddCard("card-ecological-zone")
-
-	// Place a greenery at (0,0,0)
-	greenery := shared.HexPosition{Q: 0, R: 0, S: 0}
-	err := testGame.Board().UpdateTileOccupancy(ctx, greenery,
-		board.TileOccupant{Type: shared.ResourceGreeneryTile}, p.ID())
-	testutil.AssertNoError(t, err, "placing greenery")
-
-	playCardAction := cardAction.NewPlayCardAction(repo, cardRegistry, nil, logger)
-	payment := cardAction.PaymentRequest{Credits: 12}
-	err = playCardAction.Execute(ctx, testGame.ID(), p.ID(), "card-ecological-zone", payment, nil, nil, nil, nil)
-	testutil.AssertNoError(t, err, "Ecological Zone should play successfully")
-
-	selection := testGame.GetPendingTileSelection(p.ID())
-	testutil.AssertTrue(t, selection != nil, "Should have pending tile selection")
-	testutil.AssertEqual(t, "greenery", selection.TileType, "Pending tile type should be greenery")
-
-	// All available hexes should be adjacent to the greenery at (0,0,0)
-	greeneryNeighbors := greenery.GetNeighbors()
-	neighborStrings := make(map[string]bool)
-	for _, n := range greeneryNeighbors {
-		neighborStrings[n.String()] = true
-	}
-
-	for _, hex := range selection.AvailableHexes {
-		testutil.AssertTrue(t, neighborStrings[hex],
-			"Hex "+hex+" should be adjacent to greenery at (0,0,0)")
-	}
-	testutil.AssertTrue(t, len(selection.AvailableHexes) > 0,
-		"Should have at least one available hex adjacent to greenery")
-}
-
-func TestEcologicalZone_NoGreeneryYields0AvailableHexes(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, _ := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	ctx := context.Background()
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	// No greenery placed — should have 0 available hexes
-	count := testGame.CountAvailableHexesForTile("greenery", p.ID(), &shared.TileRestrictions{
-		AdjacentToType: "greenery",
-	})
-	testutil.AssertEqual(t, 0, count, "No hexes should be available when no greenery exists")
-}
-
-// --- AdjacentToOwned restriction ---
-
-func TestAdjacentToOwned_OnlyCountsPlayerOwnedTiles(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, _ := testutil.CreateTestGameWithPlayers(t, 2, broadcaster)
-	ctx := context.Background()
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	other := players[1]
-	p.SetCorporationID("corp-tharsis-republic")
-	other.SetCorporationID("corp-mining-guild")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	// Place a greenery owned by OTHER player at (0,0,0)
-	greenery := shared.HexPosition{Q: 0, R: 0, S: 0}
-	err := testGame.Board().UpdateTileOccupancy(ctx, greenery,
-		board.TileOccupant{Type: shared.ResourceGreeneryTile}, other.ID())
-	testutil.AssertNoError(t, err, "placing greenery for other player")
-
-	// With adjacentToOwned=true, player p should have 0 available hexes
-	// because the greenery belongs to the other player
-	count := testGame.CountAvailableHexesForTile("greenery", p.ID(), &shared.TileRestrictions{
-		AdjacentToType:  "greenery",
-		AdjacentToOwned: true,
-	})
-	testutil.AssertEqual(t, 0, count, "No hexes should be available when only opponent owns greenery")
-
-	// Place a greenery owned by player p at (2,0,-2)
-	greenery2 := shared.HexPosition{Q: 2, R: 0, S: -2}
-	err = testGame.Board().UpdateTileOccupancy(ctx, greenery2,
-		board.TileOccupant{Type: shared.ResourceGreeneryTile}, p.ID())
-	testutil.AssertNoError(t, err, "placing greenery for player")
-
-	// Now player p should have some available hexes adjacent to their own greenery
-	count = testGame.CountAvailableHexesForTile("greenery", p.ID(), &shared.TileRestrictions{
-		AdjacentToType:  "greenery",
-		AdjacentToOwned: true,
-	})
-	testutil.AssertTrue(t, count > 0, "Should have available hexes adjacent to own greenery")
-}
-
-// --- State calculator: restricted tile placements produce errors, not warnings ---
-
-func makeUrbanizedAreaCard() gamecards.Card {
-	minAdj := 2
-	return gamecards.Card{
-		ID:   "card-urbanized-area",
-		Name: "Urbanized Area",
-		Type: gamecards.CardTypeAutomated,
-		Cost: 10,
-		Tags: []shared.CardTag{shared.TagBuilding, shared.TagCity},
-		Behaviors: []shared.CardBehavior{
-			{
-				Triggers: []shared.Trigger{{Type: shared.TriggerTypeAuto}},
-				Outputs: []shared.BehaviorCondition{
-					shared.NewProductionCondition(shared.ResourceCreditProduction, 2, "self-player"),
-					shared.NewProductionCondition(shared.ResourceEnergyProduction, -1, "self-player"),
-					&shared.TilePlacementCondition{
-						ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceCityPlacement, Amount: 1, Target: "none"},
-						TileRestrictions: &shared.TileRestrictions{
-							AdjacentToType:    "city",
-							MinAdjacentOfType: &minAdj,
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func TestUrbanizedArea_StateCalculatorReturnsErrorWhenNoPlacements(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, _ := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	ctx := context.Background()
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{shared.ResourceCredit: 100})
-	p.Resources().AddProduction(map[shared.ResourceType]int{shared.ResourceEnergyProduction: 1})
-
-	urbanizedArea := makeUrbanizedAreaCard()
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{urbanizedArea})
-
-	// Only 1 city on the board — Urbanized Area requires adjacent to 2 cities
-	city1 := shared.HexPosition{Q: 0, R: 0, S: 0}
-	err := testGame.Board().UpdateTileOccupancy(ctx, city1,
-		board.TileOccupant{Type: shared.ResourceCityTile}, "other-player")
-	testutil.AssertNoError(t, err, "placing city")
-
-	state := action.CalculatePlayerCardState(&urbanizedArea, p, testGame, cardRegistry)
-
-	testutil.AssertTrue(t, !state.Available(),
-		"Urbanized Area should NOT be available when only 1 city exists")
-
-	hasNoCityPlacementError := false
-	for _, e := range state.Errors {
-		if e.Code == player.ErrorCodeNoCityPlacements {
-			hasNoCityPlacementError = true
-		}
-	}
-	testutil.AssertTrue(t, hasNoCityPlacementError,
-		"Should have ErrorCodeNoCityPlacements error")
-}
-
-func TestUrbanizedArea_PlayCardRejectedWhenNoPlacements(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, repo := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	logger := testutil.TestLogger()
-	ctx := context.Background()
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{shared.ResourceCredit: 100})
-	p.Resources().AddProduction(map[shared.ResourceType]int{shared.ResourceEnergyProduction: 1})
-
-	urbanizedArea := makeUrbanizedAreaCard()
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{urbanizedArea})
-	p.Hand().AddCard("card-urbanized-area")
-
-	// Only 1 city — cannot satisfy adjacency requirement
-	city1 := shared.HexPosition{Q: 0, R: 0, S: 0}
-	err := testGame.Board().UpdateTileOccupancy(ctx, city1,
-		board.TileOccupant{Type: shared.ResourceCityTile}, "other-player")
-	testutil.AssertNoError(t, err, "placing city")
-
-	playCardAction := cardAction.NewPlayCardAction(repo, cardRegistry, nil, logger)
-	payment := cardAction.PaymentRequest{Credits: 10}
-	err = playCardAction.Execute(ctx, testGame.ID(), p.ID(), "card-urbanized-area", payment, nil, nil, nil, nil)
-
-	testutil.AssertTrue(t, err != nil,
-		"Playing Urbanized Area should fail when no valid placements exist")
-}
-
-// --- Plantation (193) ---
-// "Place a greenery tile and raise oxygen 1 step."
-// Per TM rules, greenery must be adjacent to player's own tiles (with fallback to anywhere).
-
-func makePlantationCard() gamecards.Card {
-	return gamecards.Card{
-		ID:   "card-plantation",
-		Name: "Plantation",
-		Type: gamecards.CardTypeAutomated,
-		Cost: 15,
-		Tags: []shared.CardTag{shared.TagPlant},
-		Behaviors: []shared.CardBehavior{
-			{
-				Triggers: []shared.Trigger{{Type: shared.TriggerTypeAuto}},
-				Outputs: []shared.BehaviorCondition{
-					&shared.TilePlacementCondition{
-						ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceGreeneryPlacement, Amount: 1, Target: "none"},
-					},
-				},
-			},
-		},
-	}
-}
-
-func TestPlantation_GreeneryAdjacentToOwnedTiles(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, repo := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	logger := testutil.TestLogger()
-	ctx := context.Background()
-
-	plantation := makePlantationCard()
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{plantation})
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{
-		shared.ResourceCredit: 100,
-	})
-	p.Hand().AddCard("card-plantation")
-
-	// Place a city owned by the current player at (0,0,0)
-	ownedTile := shared.HexPosition{Q: 0, R: 0, S: 0}
-	err := testGame.Board().UpdateTileOccupancy(ctx, ownedTile,
-		board.TileOccupant{Type: shared.ResourceCityTile}, p.ID())
-	testutil.AssertNoError(t, err, "placing owned city")
-
-	playCardAction := cardAction.NewPlayCardAction(repo, cardRegistry, nil, logger)
-	payment := cardAction.PaymentRequest{Credits: 15}
-	err = playCardAction.Execute(ctx, testGame.ID(), p.ID(), "card-plantation", payment, nil, nil, nil, nil)
-	testutil.AssertNoError(t, err, "Plantation should play successfully")
-
-	selection := testGame.GetPendingTileSelection(p.ID())
-	testutil.AssertTrue(t, selection != nil, "Should have pending tile selection")
-	testutil.AssertEqual(t, "greenery", selection.TileType, "Pending tile type should be greenery")
-
-	// All available hexes should be adjacent to the owned tile at (0,0,0)
-	neighbors := ownedTile.GetNeighbors()
-	neighborStrings := make(map[string]bool)
-	for _, n := range neighbors {
-		neighborStrings[n.String()] = true
-	}
-
-	for _, hex := range selection.AvailableHexes {
-		testutil.AssertTrue(t, neighborStrings[hex],
-			"Hex "+hex+" should be adjacent to owned tile at (0,0,0)")
-	}
-	testutil.AssertTrue(t, len(selection.AvailableHexes) > 0,
-		"Should have at least one available hex adjacent to owned tile")
-}
-
-func TestPlantation_FallbackWhenNoOwnedTiles(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, repo := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	logger := testutil.TestLogger()
-	ctx := context.Background()
-
-	plantation := makePlantationCard()
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{plantation})
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{
-		shared.ResourceCredit: 100,
-	})
-	p.Hand().AddCard("card-plantation")
-
-	// No tiles placed — player has no owned tiles on board
-
-	playCardAction := cardAction.NewPlayCardAction(repo, cardRegistry, nil, logger)
-	payment := cardAction.PaymentRequest{Credits: 15}
-	err := playCardAction.Execute(ctx, testGame.ID(), p.ID(), "card-plantation", payment, nil, nil, nil, nil)
-	testutil.AssertNoError(t, err, "Plantation should play successfully with fallback")
-
-	selection := testGame.GetPendingTileSelection(p.ID())
-	testutil.AssertTrue(t, selection != nil, "Should have pending tile selection")
-	testutil.AssertEqual(t, "greenery", selection.TileType, "Pending tile type should be greenery")
-
-	// With no owned tiles, fallback allows all valid land tiles
-	testutil.AssertTrue(t, len(selection.AvailableHexes) > 1,
-		"Should have multiple available hexes when no owned tiles (fallback)")
-}
-
-// --- Mangrove (059) ---
-// "Place a greenery tile ON AN AREA RESERVED FOR OCEAN and raise oxygen 1 step."
-// onTileType: "ocean" overrides normal placement — should NOT be affected by adjacency fix.
-
-func makeMangroveCard() gamecards.Card {
-	return gamecards.Card{
-		ID:   "card-mangrove",
-		Name: "Mangrove",
-		Type: gamecards.CardTypeAutomated,
-		Cost: 12,
-		Tags: []shared.CardTag{shared.TagPlant},
-		Behaviors: []shared.CardBehavior{
-			{
-				Triggers: []shared.Trigger{{Type: shared.TriggerTypeAuto}},
-				Outputs: []shared.BehaviorCondition{
-					&shared.TilePlacementCondition{
-						ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceGreeneryPlacement, Amount: 1, Target: "none"},
-						TileRestrictions: &shared.TileRestrictions{
-							OnTileType: "ocean",
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func TestMangrove_NotAffectedByAdjacentToOwned(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	testGame, repo := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	logger := testutil.TestLogger()
-	ctx := context.Background()
-
-	mangrove := makeMangroveCard()
-	cardRegistry := cards.NewInMemoryCardRegistry([]gamecards.Card{mangrove})
-
-	players := testGame.GetAllPlayers()
-	p := players[0]
-	p.SetCorporationID("corp-tharsis-republic")
-
-	testutil.AssertNoError(t, testGame.UpdateStatus(ctx, shared.GameStatusActive), "update status")
-	testutil.AssertNoError(t, testGame.UpdatePhase(ctx, shared.GamePhaseAction), "update phase")
-	testutil.AssertNoError(t, testGame.SetCurrentTurn(ctx, p.ID(), 2), "set current turn")
-
-	p.Resources().Add(map[shared.ResourceType]int{
-		shared.ResourceCredit: 100,
-	})
-	p.Hand().AddCard("card-mangrove")
-
-	// Place a city far away — Mangrove should still offer ocean spaces, not adjacent-to-owned
-	ownedTile := shared.HexPosition{Q: 4, R: -2, S: -2}
-	err := testGame.Board().UpdateTileOccupancy(ctx, ownedTile,
-		board.TileOccupant{Type: shared.ResourceCityTile}, p.ID())
-	testutil.AssertNoError(t, err, "placing owned city")
-
-	playCardAction := cardAction.NewPlayCardAction(repo, cardRegistry, nil, logger)
-	payment := cardAction.PaymentRequest{Credits: 12}
-	err = playCardAction.Execute(ctx, testGame.ID(), p.ID(), "card-mangrove", payment, nil, nil, nil, nil)
-	testutil.AssertNoError(t, err, "Mangrove should play successfully")
-
-	selection := testGame.GetPendingTileSelection(p.ID())
-	testutil.AssertTrue(t, selection != nil, "Should have pending tile selection")
-	testutil.AssertEqual(t, "greenery", selection.TileType, "Pending tile type should be greenery")
-
-	// Mangrove should offer ocean spaces — verify available hexes are ocean tiles
-	testutil.AssertTrue(t, len(selection.AvailableHexes) > 0,
-		"Should have available ocean hexes for Mangrove")
-
-	// Verify that hexes include non-adjacent-to-owned ocean spaces
-	ownedNeighbors := ownedTile.GetNeighbors()
-	ownedNeighborStrings := make(map[string]bool)
-	for _, n := range ownedNeighbors {
-		ownedNeighborStrings[n.String()] = true
-	}
-
-	hasNonAdjacentHex := false
-	for _, hex := range selection.AvailableHexes {
-		if !ownedNeighborStrings[hex] {
-			hasNonAdjacentHex = true
-			break
-		}
-	}
-	testutil.AssertTrue(t, hasNonAdjacentHex,
-		"Mangrove should offer ocean spaces not adjacent to owned tiles")
-}
