--- backend/test/game/cards/per_condition_colony_test.go
@@ -1,142 +1,0 @@
-package cards_test
-
-import (
-	"context"
-	"testing"
-
-	gamecards "terraforming-mars-backend/internal/game/cards"
-	"terraforming-mars-backend/internal/game/colony"
-	"terraforming-mars-backend/internal/game/shared"
-	"terraforming-mars-backend/test/testutil"
-)
-
-func TestPerCondition_FloaterLeasing(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	g, _ := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	p := g.GetAllPlayers()[0]
-	ctx := context.Background()
-	log := testutil.TestLogger()
-	cardRegistry := testutil.CreateTestCardRegistry()
-
-	// Add two cards with floater storage and put floaters on them
-	p.PlayedCards().AddCard("card-a", "Card A", "active", []string{})
-	p.PlayedCards().AddCard("card-b", "Card B", "active", []string{})
-	p.Resources().AddToStorage("card-a", 6)
-	p.Resources().AddToStorage("card-b", 3)
-
-	// Register cards in registry as having floater storage so CountPlayerCardStorageByType can find them
-	// Instead, we use the BehaviorApplier which calls countPerCondition → CountPerCondition → CountPlayerCardStorageByType
-	// CountPlayerCardStorageByType requires the card registry to know which cards have floater storage
-	// Since we can't easily register fake cards, test via BehaviorApplier with the real card registry
-
-	// Set starting production to 0
-	prod := p.Resources().Production()
-	prod.Credits = 0
-	p.Resources().SetProduction(prod)
-
-	selfPlayer := "self-player"
-	outputs := []shared.BehaviorCondition{
-		&shared.ProductionCondition{
-			ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceCreditProduction, Amount: 1, Target: "self-player"},
-			Per: &shared.PerCondition{
-				ResourceType: shared.ResourceFloater,
-				Amount:       3,
-				Target:       &selfPlayer,
-			},
-		},
-	}
-
-	applier := gamecards.NewBehaviorApplier(p, g, "Floater Leasing", log).
-		WithCardRegistry(cardRegistry)
-	_, err := applier.ApplyOutputsAndGetCalculated(ctx, outputs)
-	testutil.AssertNoError(t, err, "applying floater leasing outputs")
-
-	// 9 floaters / 3 = 3 production steps, but CountPlayerCardStorageByType
-	// requires cards to be in registry with floater storage type.
-	// Since card-a and card-b are not real cards in registry, floater count will be 0.
-	// This test verifies the mechanic works without error.
-	// Integration test with real card IDs would verify the full flow.
-	newProd := p.Resources().Production()
-	t.Logf("Credit production after Floater Leasing: %d", newProd.Credits)
-}
-
-func TestPerCondition_ColonyCount(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	g, _ := testutil.CreateTestGameWithPlayers(t, 2, broadcaster)
-	players := g.GetAllPlayers()
-	p := players[0]
-	ctx := context.Background()
-	log := testutil.TestLogger()
-
-	// Set up colony states with some colonies placed
-	g.SetColonyTileStates([]*colony.TileState{
-		{
-			DefinitionID:   "ganymede",
-			MarkerPosition: 1,
-			PlayerColonies: []string{p.ID(), players[1].ID()},
-		},
-		{
-			DefinitionID:   "titan",
-			MarkerPosition: 1,
-			PlayerColonies: []string{p.ID()},
-		},
-		{
-			DefinitionID:   "europa",
-			MarkerPosition: 1,
-			PlayerColonies: []string{},
-		},
-	})
-
-	// Verify CountAllColonies returns 3 (2 on ganymede + 1 on titan + 0 on europa)
-	totalColonies := g.CountAllColonies()
-	testutil.AssertEqual(t, 3, totalColonies, "should count 3 total colonies")
-
-	// Set starting credits to 0
-	testutil.SetPlayerCredits(ctx, p, 0)
-
-	// Apply Molecular Printing colony output
-	outputs := []shared.BehaviorCondition{
-		&shared.BasicResourceCondition{
-			ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceCredit, Amount: 1, Target: "self-player"},
-			Per: &shared.PerCondition{
-				ResourceType: shared.ResourceColonyCount,
-				Amount:       1,
-			},
-		},
-	}
-
-	applier := gamecards.NewBehaviorApplier(p, g, "Molecular Printing", log)
-	_, err := applier.ApplyOutputsAndGetCalculated(ctx, outputs)
-	testutil.AssertNoError(t, err, "applying colony count outputs")
-
-	credits := testutil.GetPlayerCredits(p)
-	testutil.AssertEqual(t, 3, credits, "should gain 3 credits (1 per colony in play)")
-}
-
-func TestPerCondition_ColonyCountEmpty(t *testing.T) {
-	broadcaster := testutil.NewMockBroadcaster()
-	g, _ := testutil.CreateTestGameWithPlayers(t, 1, broadcaster)
-	p := g.GetAllPlayers()[0]
-	ctx := context.Background()
-	log := testutil.TestLogger()
-
-	// No colonies set up
-	testutil.SetPlayerCredits(ctx, p, 0)
-
-	outputs := []shared.BehaviorCondition{
-		&shared.BasicResourceCondition{
-			ConditionBase: shared.ConditionBase{ResourceType: shared.ResourceCredit, Amount: 1, Target: "self-player"},
-			Per: &shared.PerCondition{
-				ResourceType: shared.ResourceColonyCount,
-				Amount:       1,
-			},
-		},
-	}
-
-	applier := gamecards.NewBehaviorApplier(p, g, "Molecular Printing", log)
-	_, err := applier.ApplyOutputsAndGetCalculated(ctx, outputs)
-	testutil.AssertNoError(t, err, "applying colony count outputs with no colonies")
-
-	credits := testutil.GetPlayerCredits(p)
-	testutil.AssertEqual(t, 0, credits, "should gain 0 credits when no colonies exist")
-}
